1) Донастроил парсер 2) Добавил две ручки для получения страницы клуба 3)Донастроил модели 4) Добавил в сервис кастомный принт

+++ src/club/models.py
@@ -2,11 +2,12 @@ from django.db import models
 
 
 class Club(models.Model):
-    identifier = models.PositiveSmallIntegerField(db_index=True, unique=True)
+    identifier = models.PositiveIntegerField(db_index=True, unique=True)
     name = models.CharField(max_length=63)
     photo = models.FileField(upload_to=".../media/clubs", default=None, null=True, blank=True)
     photo_url = models.TextField(default=None, null=True, blank=True)
     stadium_name = models.CharField(max_length=63)
     stadium_count_of_sits = models.PositiveIntegerField(default=0)
     city_name = models.CharField(max_length=63)
+    # расстановка команды на поле
     placement = models.JSONField(default=dict)
